# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InfoSec Harness",
    "version": "2.0.0.dev0"
  },
  "paths": {
...
    "/api/batches": {
      "post": {
        "summary": "Submit",
        "operationId": "submit_api_batches_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Batches",
        "operationId": "batches_api_batches_get",
        "parameters": [
          {
            "name": "population",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "operational",
                    "demo",
                    "legacy"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Population"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchSummary"
                  },
                  "title": "Response Batches Api Batches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/batches/ado": {
      "post": {
        "summary": "Submit Ado",
        "operationId": "submit_ado_api_batches_ado_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitADORequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdoBatchAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/batches/{batch_id}": {
      "get": {
        "summary": "Batch",
        "operationId": "batch_api_batches__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "population",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "operational",
                    "demo",
                    "legacy"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Population"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "get": {
        "summary": "Runs",
        "operationId": "runs_api_runs_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Batch Id"
            }
          },
          {
            "name": "verdict",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verdict"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "population",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "operational",
                    "demo",
                    "legacy"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Population"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response Runs Api Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "summary": "Run",
        "operationId": "run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "population",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "operational",
                    "demo",
                    "legacy"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Population"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/review": {
      "post": {
        "summary": "Review",
        "operationId": "review_api_runs__run_id__review_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSaved"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/experiments": {
      "get": {
        "summary": "Experiments",
        "operationId": "experiments_api_experiments_get",
        "parameters": [
          {
            "name": "population",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "operational",
                    "demo",
                    "legacy"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Population"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExperimentSummary"
                  },
                  "title": "Response Experiments Api Experiments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/batches/{batch_id}/cancel": {
      "post": {
        "summary": "Cancel Batch",
        "operationId": "cancel_batch_api_batches__batch_id__cancel_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdoBatchAccepted": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "imported": {
            "type": "integer",
            "title": "Imported"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "imported"
        ],
        "title": "AdoBatchAccepted"
      },
...
      "BatchAccepted": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          }
        },
        "type": "object",
        "required": [
          "batch_id"
        ],
        "title": "BatchAccepted"
      },
      "BatchDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/BatchStatus"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "finding_count": {
            "type": "integer",
            "title": "Finding Count"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Status Counts"
          },
          "current_phases": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Current Phases"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "last_activity_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Activity At"
          },
          "budget": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget"
          },
          "verdict_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Verdict Counts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "label",
          "source_kind",
          "finding_count",
          "created_at",
          "status_counts",
          "budget",
          "verdict_counts"
        ],
        "title": "BatchDetail"
      },
...
      "BatchSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/BatchStatus"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "finding_count": {
            "type": "integer",
            "title": "Finding Count"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Status Counts"
          },
          "current_phases": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Current Phases"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "last_activity_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Activity At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "label",
          "source_kind",
          "finding_count",
          "created_at"
        ],
        "title": "BatchSummary"
      },
...
      "CancelResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/BatchStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "CancelResponse"
      },
...
      "ExperimentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExperimentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "dataset": {
            "type": "string",
            "title": "Dataset"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version"
          },
          "git_sha": {
            "type": "string",
            "title": "Git Sha"
          },
          "overlay": {
            "type": "string",
            "title": "Overlay"
          },
          "repetitions": {
            "type": "integer",
            "title": "Repetitions"
          },
          "metrics": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Metrics"
          },
          "config_hash": {
            "type": "string",
            "title": "Config Hash"
          },
          "git_dirty": {
            "type": "boolean",
            "title": "Git Dirty"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "backend": {
            "type": "string",
            "title": "Backend"
          },
          "pricing": {
            "type": "string",
            "title": "Pricing"
          },
          "harness_version": {
            "type": "string",
            "title": "Harness Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agent",
          "status",
          "dataset",
          "dataset_version",
          "git_sha",
          "overlay",
          "repetitions",
          "metrics",
          "config_hash",
          "git_dirty",
          "model_name",
          "backend",
          "pricing",
          "harness_version",
          "created_at"
        ],
        "title": "ExperimentSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {},
...
      "ReviewRequest": {
        "properties": {
          "reviewer": {
            "type": "string",
            "title": "Reviewer",
            "default": ""
          },
          "decision": {
            "type": "string",
            "enum": [
              "confirm",
              "override"
            ],
            "title": "Decision"
          },
          "override_label": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerdictLabel"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "ReviewRequest"
      },
      "ReviewSaved": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ReviewSaved"
      },
...
      "RunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "fingerprint": {
            "type": "string",
            "title": "Fingerprint"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "repo_url": {
            "type": "string",
            "title": "Repo Url"
          },
          "revision": {
            "type": "string",
            "title": "Revision"
          },
          "cwe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwe"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "inconclusive_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inconclusive Reason"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "priority_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Score"
          },
          "environment_scope": {
            "type": "string",
            "title": "Environment Scope"
          },
          "early_exit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Early Exit"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens"
          },
          "cache_read_tokens": {
            "type": "integer",
            "title": "Cache Read Tokens"
          },
          "latency_s": {
            "type": "number",
            "title": "Latency S"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "telemetry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunTelemetry"
              },
              {
                "type": "null"
              }
            ]
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_id",
          "fingerprint",
          "title",
          "repo_url",
          "revision",
          "cwe",
          "severity",
          "status",
          "verdict",
          "confidence",
          "inconclusive_reason",
          "priority",
          "priority_score",
          "environment_scope",
          "early_exit",
          "cost_usd",
          "total_tokens",
          "cache_read_tokens",
          "latency_s",
          "created_at",
          "telemetry",
          "phase"
        ],
        "title": "RunSummary"
      },
...
      "SubmitADORequest": {
        "properties": {
          "work_item_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Work Item Ids"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "work_item_ids"
        ],
        "title": "SubmitADORequest"
      },
      "SubmitRequest": {
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingInput"
            },
            "type": "array",
            "title": "Findings"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": ""
          },
          "mode": {
            "type": "string",
            "enum": [
              "local",
              "temporal"
            ],
            "title": "Mode",
            "default": "temporal"
          }
        },
        "type": "object",
        "required": [
          "findings"
        ],
        "title": "SubmitRequest"
      },
...
      "VerdictLabel": {
        "type": "string",
        "enum": [
          "potentially_exploitable",
          "likely_not_exploitable",
          "inconclusive"
        ],
        "title": "VerdictLabel"
      }
    }
  }
}